+++ src/madness/mra/macrotaskpartitioner.h
@@ -105,7 +105,7 @@ public:
     /// given vector v, copy vector elements of v_batch into vector
     template<typename vecT>
     vecT insert_batch(vecT v, const vecT& v_batch) const {
-        MADNESS_CHECK(v_batch.size()==this->size() or this->is_full_size());
+      MADNESS_CHECK(v_batch.size()==size_t(this->size()) or this->is_full_size());
         std::copy(v_batch.begin(), v_batch.end(), v.begin()+begin);
         return v;
     }
@@ -261,7 +261,7 @@ public:
         if (policy == "guided") {
             long begin = 0;
             long end = 0;
-            while (end < vsize) {
+            while (end < long(vsize)) {
                 end += std::min(max_batch_size, std::max(min_batch_size, ((vsize - end) / nsubworld)));
                 end = std::min(end, long(vsize));
                 Batch batch(Batch_1D(begin, end),Batch_1D(begin,end));
